Add saved teams to team select; force team names to "Saved Team X"

+++ team-setup-model.ts
@@ -15,8 +15,8 @@
         };
         let arr1 = team1.lineup.map(map);
         let arr2 = team2.lineup.map(map);
-        this.json1 = ko.observable(JSON.stringify({ starters: arr1 }, undefined, 4));
-        this.json2 = ko.observable(JSON.stringify({ starters: arr2 }, undefined, 4));
+        this.json1 = ko.observable(JSON.stringify(arr1, undefined, 4));
+        this.json2 = ko.observable(JSON.stringify(arr2, undefined, 4));
     }
 
     swap() {
@@ -32,16 +32,21 @@
     reset() {
         reloadViewModel(team1, team2);
         this.onClose();
-        if (confirm("Would you also like to clear the previously set custom teams so they don't load next time?")) {
+        if (window.localStorage && confirm("Would you also like to clear the previously set custom teams so they don't load next time?")) {
             localStorage.removeItem("team1");
             localStorage.removeItem("team2");
         }
     }
 
     save() {
-        localStorage.setItem("team1", this.json1());
-        localStorage.setItem("team2", this.json2());
-        reloadViewModel(JSON.parse(this.json1()), JSON.parse(this.json2()));
+        if (window.localStorage) {
+            localStorage.setItem("team1", this.json1());
+            localStorage.setItem("team2", this.json2());
+        }
+        reloadViewModel(
+            { name: "Saved Team 1", starters: JSON.parse(this.json1()) },
+            { name: "Saved Team 2", starters: JSON.parse(this.json2()) }
+        );
         this.onClose();
     }
 }
